First Draft of Route & Schedule flow
Added a feature that allows to add multiple new routes in "my routes" through modal input. The latest route is then visible in the schedule section. This code is not stable yet and routes are lost when the app is refreshed. Furthermore, I haven't figured out how to store multiple routes yet.